Added in Enemies
converted Orbiter.as, Enemy.as, and EnemyController.as. Tweens do not exist in Gamemaker so used Gamemaker function motion_set along with some polar math to get the angle and the speed of the enemies in the orbiting movement. Added collision between player and enemies. Currently enemies spawn with a button press for testing.